@sbourdeauducq 2021-03-18
I would just make it a hard dependency. Otherwise it'll just create inconsistent performance results, and make the code a bit more complicated than it could be.