@sbourdeauducq 2021-03-18
There's no py35 conda package for pybase64, so artiq-legacy will have to keep using the older sipyco version. How is your nix-fu?