@airwoodix 2021-06-09
> I would just make it a hard dependency.

Fixed in 5f05040.

> There's no py35 conda package for pybase64, so artiq-legacy will have to keep using the older sipyco version.

Is this still an issue? (Python 3.5 is anyway EOL)