@sbourdeauducq 2021-06-09
> Is this still an issue?

Definitely. Many labs are still using ARTIQ-5.

> (Python 3.5 is anyway EOL)

Yes, just like Windows is a shitty OS, but most people are still using it.