@airwoodix 2021-06-09
Would it be an option to make an ad-hoc package, like for ``qasync``, ``pythonparser`` and others? Or was my quick scan of ``nix-scripts`` too fast?
Otherwise, if packaging is too annoying, one could also skip this part of the patch as it only brings 10% extra (but with negligible complexity in the code).