@sbourdeauducq 2021-06-09
It is an option, but that's not easier than keeping the older sipyco version on the stable channels.