@airwoodix 2021-06-15
> How is your nix-fu?

I'm not quite sure what would need to be done, so I'd say "not good enough".

I'd nevertheless be very happy to see this patch merged. How would you like to proceed?