@sbourdeauducq 2021-06-16
This patch when merged will break the CI for Windows and possibly Windows user installations. Sorry I do not have time currently to fix it for you.