@sbourdeauducq 2021-06-16
It's merged, but there will be conda problems when the package source is updated.